Final commit for CRUD operation

+++ Backend/src/main/java/org/launchCode/procrastiNOT/controllers/api/ApiTaskController.java
@@ -12,28 +12,19 @@ import java.util.Optional;
 
 @CrossOrigin(origins="http://localhost:3000", maxAge = 3600)
 @RestController
-@RequestMapping("/api/task")
+@RequestMapping("/login/task")
+
 public class ApiTaskController {
     @Autowired
     private TaskRepository taskRepository;// injecting  CRUDRepository
 
-//    @PostMapping("/")
-//    Task newTask(@RequestBody Task newTask) {
-//
-//        return taskRepository.save(newTask);
-//
-//    }
-
-    @PostMapping("/")
+    @PostMapping("/add")
     public ResponseEntity postTheTask(@RequestBody Task newTask) {
-//        return Event.createEvent(eventDto.getName(), eventDto.getDescription());
         return new ResponseEntity<>(taskRepository.save(newTask), HttpStatus.CREATED);
     }
 
-    @GetMapping("/")
-//    List<Task> getAllTask() {
+    @GetMapping("/alltasks")
     public ResponseEntity getAllTasks() {
-
         List<Task> tasks = (List<Task>) taskRepository.findAll();
         return new ResponseEntity<>(tasks, HttpStatus.OK);
 
@@ -43,12 +34,10 @@ public class ApiTaskController {
     @GetMapping("/{id}")//ResponseEntity is class
     public ResponseEntity getTaskById(@PathVariable Integer id) {
 
-        //  if (taskId !=null) {
-//        List<Task> optTask = taskRepository.findById(taskId);
+
         Optional<Task> optTask = taskRepository.findById(id);
 
         if (optTask.isPresent()) {
-//            Task task = (Task) optTask.get(taskId);// Get task by ID
             return new ResponseEntity<>(optTask.get(), HttpStatus.OK);
         } else {
             return new ResponseEntity<>(optTask.get(), HttpStatus.NOT_FOUND);
@@ -58,30 +47,42 @@ public class ApiTaskController {
     @PutMapping("/{id}")
     public ResponseEntity updateTask(@PathVariable int id, @RequestBody Task newTask) {
         Optional<Task> optTask = taskRepository.findById(id);
-//        List<Task> optTask = taskRepository.findById(id);
-//        Task theTask = taskRepository.findById(id);
-
-////        List<Task> optTask = taskRepository.findById(Id);
-//      if (optTask.isPresent()) {
-//        if (id != null) {
-
-
-            if (!taskRepository.existsById(id)) {
-//                optTask.set(getTaskById(id));
-                return new ResponseEntity(taskRepository.save(newTask), HttpStatus.ACCEPTED);
-        }
-
-
-        return null;
+        if (optTask.isPresent()) {
+            Task task = (Task) optTask.get();
+            newTask.setDescription(newTask.getDescription());
+            return new ResponseEntity(taskRepository.save(newTask), HttpStatus.ACCEPTED);
+        } else if (!taskRepository.existsById(id)) {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        } else
+            return null;
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity deleteTheTask(@PathVariable int id) {
-//        boolean deleted = Event.deleteItem(id);
-        taskRepository.deleteById(id);
-        if (!taskRepository.existsById(id)) {
-            return new ResponseEntity(HttpStatus.NO_CONTENT);
-        }
-        return new ResponseEntity(HttpStatus.NOT_FOUND);
+
+//    @DeleteMapping("/{id}")
+//    public ResponseEntity deleteTheTask(@PathVariable int id) {
+//        taskRepository.deleteById(id);
+//        if (!taskRepository.existsById(id)) {
+//            return new ResponseEntity(HttpStatus.NO_CONTENT);
+//        }
+//        return new ResponseEntity(HttpStatus.NOT_FOUND);
+//    }
+
+
+
+   @DeleteMapping("/{ids}")
+    public ResponseEntity<?> deleteByIds(@PathVariable("ids") List<Integer> ids){
+       ids.forEach(id->{
+           if(taskRepository.existsById(id)){
+
+               taskRepository.deleteById(id);
+            }
+        });
+        return ResponseEntity.ok().body("Task removed");
     }
+
+
 }
+
+
+
+
